Update to new upstream release 3.8.3.
  Atlas is a package that requires being compiled in the computer
where it will be used. To achieve this, a "ready to build" setup
is available in the main package, and generic precompiled library
binaries are also available. The default ones are:
o libatlas-sse2 for ix86
  This requires a Pentium IV class cpu ("modern" ix86 cpus support
  sse3, but Celeron and older Pentium IV class cpus support only sse2).
o libatlas-x86_64 for x86_64
  Generic x86_64 build.

  Note that the build is dependent on the computer generating the
packages, and the threaded/smp libraries will only be generated on
builds on a multi core/cpu computer only.


SILENT: svn-revision: 350249